Move type logic to EventTypes module
